fix(buildsource): don't credit degraded-pass edges as coverage, split inline.py's graph folding
A full pass that hit per-TU diagnostics correctly never sets extractor_passes
(seventh review's rule), but still folds edges from the TUs that did parse.
Nothing recorded this, so those surviving edges fell into the per-kind
fallback's weak "bare edge presence" trust tier, letting a degraded
baseline's edge be compared against a clean candidate's edge in a wholly
different, never-successfully-parsed TU.

Add SourceGraphSummary.degraded_passes: dict[str, bool] (additive, same
round-trip pattern as narrowed_passes/narrowed_scope), set whenever a pass
examined units but extractor.diagnostics was non-empty. Stamped by
inline._fold_call_graph/_fold_type_graph and
cli_buildsource_helpers._collect_call_graph. _common_dependency_edge_kinds's
old_present guard now also requires not old_degraded.

Also splits _scope_narrowed_target/_fold_call_graph/_fold_type_graph out of
inline.py (sitting at its 2000-line hard cap) into a new sibling module,
inline_graph_fold.py, so future rounds have headroom instead of re-shaving
comment lines each time.
