+++ abicheck/buildsource/CLAUDE.md
@@ -46,7 +46,8 @@ L3/L4/L5 are ordinary `ChangeKind` entries that default to `API_BREAK_KINDS`
 | `pack.py` | `BuildSourcePack` — on-disk layout + inline embedding (`to_embedded_dict`), content addressing, write/load, `to_ref()` | 028 D1/D4 |
 | `inputs_pack.py` | Flow-2 build-emitted facts protocol (**ingest** side): `InputsManifest` + `is_inputs_pack()`/`load_inputs_manifest()`/`read_source_facts()` and `ingest_inputs_pack()` — fold a build-dropped `abicheck_inputs/` dir (normalized `source_facts/*.jsonl` → L4, `build/compile_commands.json` → L3, L5 graph) into an embeddable `BuildSourcePack` **without re-running a frontend** (pure parsing). Rides the existing `merge` flow (a dir input is auto-detected/ingested in `cli_buildsource_helpers._merge_load_snapshots`). Guards third-party packs: pack-root path constraint (`_safe_pack_path`, symlink-escape + non-regular-file safe), per-record skip-with-diagnostic. `raw_ast/`/`pp/`/`deps/` are forensic only, never ingested | 035 D5 (G19.4) |
 | `inputs_emit.py` | Flow-2 **producer** side (inverse of `inputs_pack`): `init_inputs_pack()`/`append_source_facts()` (incremental per-TU) + `write_inputs_pack()` (one-shot) write a conformant `abicheck_inputs/` pack that round-trips through `ingest_inputs_pack`. Pure I/O. Used by the `abicheck-cc` wrapper (`../cc_wrapper.py`) and the Clang plugin (`contrib/abicheck-clang-plugin/`) | 035 D5 (G19.4) |
-| `inline.py` | Source-tree-centric inline collection for `dump --sources`/`--build-info`: `collect_inline_pack()` (resolve compile DB → L3, replay → L4, fold → L5), `BuildConfig`/`load_build_config()`/`discover_build_config()` (`.abicheck.yml` `build:`/`sources:`), `is_pack_dir()`. Compile-DB resolution (`_resolve_compile_db`) ends in the **zero-config inferred query** (no `--allow-build-query` flag): explicit `--build-info` → trusted `--config`/`--build-query` command → `build.compile_db` glob → auto-discovered `compile_commands.json` → inferred build-system query (`build_query.py`). An arbitrary `build.query` from an auto-discovered (untrusted) `.abicheck.yml` is still never auto-run (needs explicit `--config`). | 028–033 amendment, 032 (amended) |
+| `inline.py` | Source-tree-centric inline collection for `dump --sources`/`--build-info`: `collect_inline_pack()` (resolve compile DB → L3, replay → L4, fold → L5), `BuildConfig`/`load_build_config()`/`discover_build_config()` (`.abicheck.yml` `build:`/`sources:`), `is_pack_dir()`. Compile-DB resolution (`_resolve_compile_db`) ends in the **zero-config inferred query** (no `--allow-build-query` flag): explicit `--build-info` → trusted `--config`/`--build-query` command → `build.compile_db` glob → auto-discovered `compile_commands.json` → inferred build-system query (`build_query.py`). An arbitrary `build.query` from an auto-discovered (untrusted) `.abicheck.yml` is still never auto-run (needs explicit `--config`). Call/type-graph folding lives in `inline_graph_fold.py` (split out to stay under the 2000-line cap) | 028–033 amendment, 032 (amended) |
+| `inline_graph_fold.py` | `fold_call_graph()`/`fold_type_graph()` — the Clang call/type-graph augmentation `inline._build_inline_graph` runs when `with_call_graph` is set, split out of `inline.py` to keep that file under its line-count cap. Shared `_scope_narrowed_target()` picks the scoping precedence (`changed_paths` → `scoped_units` → broad) both passes use, and stamps `extractor_passes`/`narrowed_passes`+`narrowed_scope`/`degraded_passes` on the resulting `SourceGraphSummary` per the coverage-honesty chain (ADR-041 P0, eleventh–sixteenth Codex reviews) | 041 P0 |
 | `build_query.py` | **Zero-config build-system inference** (ADR-032 amended): `detect_build_system()` (CMakeLists.txt/MODULE.bazel/makefile markers) + `inferred_query_command()` (the fixed, abicheck-authored `cmake -DCMAKE_EXPORT_COMPILE_COMMANDS=ON` / `bazel aquery` / `make -B -n -k -w` argv — never taken from the tree) + `run_inferred_build_query()` (runs it and merges CMake compile DB, Bazel aquery, or reduced-confidence Make transcript evidence into `BuildEvidence`). So `--sources` alone yields L3 with no flag and no manual compile DB step. Make evidence is transcript-scraped, so prefer a real compile DB when available. Pure detection/command-construction; only the runner touches subprocess. | 032 (amended) |
 | `pattern_scan.py` | Compiler-free lexical ABI-risk pre-scan: `scan_text()`/`scan_files()` (pure stdlib-regex over the D2 construct list), `PatternFact`/`PatternScanResult` (advisory facts + per-kind `EscalationTrigger`s feeding D7 POI focusing); never authoritative | 035 D2 (G19.1) |
 | `preprocessor_scan.py` | S2 conditional pre-scan (ADR-035 D2): `run_preprocessor_scan(build, public_headers)` runs **only** when L3 build evidence + a preprocessor (`clang -E`) are present (else returns `ran=False` + a `skipped_reason` coverage row — never clean). Pure core (`parse_defined_macros`, `select_abi_macros`/`is_abi_macro`, `find_macro_divergence`, `classify_include`/`find_private_header_leaks`) + live `ClangPreprocessorExtractor` (`-E -dM` macro capture, `-M` per-header include capture; integration-only). Emits advisory ABI-macro-value-divergence + private/generated-header-leak facts feeding D3/D4; never a verdict | 035 D2 (G19.1) |
