+++ examples/case147_scan_depth_ladder/snapshot.abi.json
@@ -19,6 +19,7 @@
     },
     "source_graph": {
       "coverage": {},
+      "degraded_passes": {},
       "edges": [],
       "external_graph_refs": [],
       "extractor_passes": {},
