+++ examples/case160_public_api_internal_dep_added/new.json
@@ -1,5 +1,6 @@
 {
   "coverage": {},
+  "degraded_passes": {},
   "edges": [
     {
       "attrs": {},
