+++ examples/case161_target_dependency_added/new.json
@@ -1,5 +1,6 @@
 {
   "coverage": {},
+  "degraded_passes": {},
   "edges": [
     {
       "attrs": {},
