+++ examples/case161_target_dependency_added/old.json
@@ -1,5 +1,6 @@
 {
   "coverage": {},
+  "degraded_passes": {},
   "edges": [],
   "external_graph_refs": [],
   "extractor_passes": {},
