+++ examples/case162_symbol_source_owner_changed/old.json
@@ -1,5 +1,6 @@
 {
   "coverage": {},
+  "degraded_passes": {},
   "edges": [
     {
       "attrs": {},
